+++ src/main/kotlin/com/daveme/chocolateCakePHP/ViewNavigationPopup.kt
@@ -1,18 +1,26 @@
 package com.daveme.chocolateCakePHP
 
 import com.daveme.chocolateCakePHP.cake.AllViewPaths
-import com.daveme.chocolateCakePHP.cake.CakePhpNavigationPresentationProvider
+import com.daveme.chocolateCakePHP.cake.CakeIcons
 import com.daveme.chocolateCakePHP.controller.CreateViewFileAction
 import com.intellij.codeInsight.navigation.PsiTargetNavigator
+import com.intellij.codeInsight.navigation.impl.PsiTargetPresentationRenderer
 import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
+import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.ui.awt.RelativePoint
+import com.jetbrains.php.PhpIcons
+import com.jetbrains.php.PhpPresentationUtil
+import com.jetbrains.php.lang.psi.elements.Method
+import com.jetbrains.php.lang.psi.elements.MethodReference
 import java.awt.Point
 import java.awt.event.MouseEvent
+import javax.swing.Icon
 import javax.swing.SwingUtilities
+
 
 fun MouseEvent.getScreenPoint(): Point? {
     // Convert the point relative to the component to a screen point
@@ -113,4 +121,50 @@ fun createViewActionPopupFromAllViewPaths(
         )
     }
     return defaultActionGroup
+}
+
+class CakePhpNavigationPresentationProvider : PsiTargetPresentationRenderer<PsiElement>() {
+    override fun getContainerText(element: PsiElement): String? {
+        val file = element.containingFile
+        if (file != null) {
+            val virtualFile = file.virtualFile
+            if (virtualFile != null)
+                return PhpPresentationUtil.getPresentablePathForFile(
+                    virtualFile,
+                    element.project
+                ) //virtualFile.presentableName
+        }
+        return super.getContainerText(element)
+    }
+
+    override fun getElementText(element: PsiElement): String {
+        val file = element.containingFile
+        if (file != null) {
+            val virtualFile = file.virtualFile
+            if (virtualFile != null) {
+                val path = virtualFile.path
+                if (path.contains("/Controller/")) {
+                    // Get containing method if call is inside a controller:
+                    val method = PsiTreeUtil.getParentOfType(element, Method::class.java)
+                    if (method != null) {
+                        return super.getElementText(method)
+                    }
+                } else if (element is MethodReference) {
+                    return virtualFile.name
+                }
+            }
+        }
+        return super.getElementText(element)
+    }
+
+    override fun getIcon(element: PsiElement): Icon {
+        val path = element.containingFile.virtualFile?.path
+        return if (path == null) {
+            PhpIcons.FUNCTION
+        } else if (path.contains("/Controller/")) {
+            CakeIcons.LOGO_SVG
+        } else {
+            PhpIcons.PHP_FILE
+        }
+    }
 }
